pyproject.toml
[tool.poetry]
name = "ylab-project"
version = "0.1.0"
description = ""
authors = ["Sergey Kim <<sergkim7@gmail.com>>"]
readme = "README.md"
packages = [{include = "restaurant_menu_app"}]

[tool.poetry.dependencies]
python = "^3.8.1"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
sqlalchemy = "^1.4.46"
alembic = "^1.9.1"
python-dotenv = "^0.21.0"
pytest = "^7.2.1"
httpx = "^0.23.3"
sqlalchemy-utils = "^0.39.0"
asyncpg = "^0.27.0"
aioredis = "^2.0.1"
pytest-asyncio = "^0.20.3"
aiofiles = "^22.1.0"
celery = "^5.2.7"
openpyxl = "^3.1.0"


[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
flake8 = "^6.0.0"
pre-commit = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
